Move Access Level check to be global
